Refactor PPO
Refactor the ppo algorithm so that it follows the following logic:
 1. Collect trajectories of length T.
 2. Iterate the collected dataset and perform K updates.

Move away from the notion of epochs. Changing the batch size while
keeping the epochs fixed will also silently result in changing the
number of updates. This behaviour is unwanted.

Better: allow the caller to set the number of updates and the batch
size and calculate the epochs as a consequence.